Fix tests for renamed files
