+++ src/containers/index.js
@@ -1,5 +1,4 @@
 export { default as App } from './App';
-export { default as LogoutPage } from './LogoutPage';
 export { default as SessionMenu } from './SessionMenu';
 export { default as StageMenu } from './StageMenu';
 export { default as Stage } from './Stage';
