+++ src/ducks/modules/__tests__/network.test.js
