+++ src/ducks/modules/__tests__/prompt.test.js
